Desktop dropdown closes as soon as mouse moves onto dropdown
Because hover was hooked to .has-dropdown > a, as soon as mouse moved away from that timer started and closed the menu. I now use the mobile detect as the main if statement, in which case .has-dropdown > a is targeted, otherwise use .has-dropdown, although the click event is hooked only to .has-dropdown > a so clicking an item doesn't close the menu on desktop.