Condensed and reverted menu backend
Condensed my menu jQuery back to much of it's original form, only retaining the function optimizations that I gathered along the way. For now, the menu operates solely with click functions, and in testing it worked just fine. It may stay the way it is, as without touch events it works on any device's input and it completely keyboard accessible.